requirements.txt
python-telegram-bot[webhooks]
gspread
oauth2client
pdfplumber

